2Tier-EC2WebApp/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary
python-dotenv==1.0.0
gunicorn
boto3
